modules reference without path now working
Involved these steps.
1 add lib.rs and move all the "mod" statements there
2 reference with "solitaire::modname"
3 make module declarations public

The name solitaire exists because this is the default name for
the library, which is taken from the name of the package.

Planning to have several performance measuring tests.
Want to be able to go back in time and compare for regressions.
So set this up with bash script to make new date stamped log file.
There can be other kinds of logs as well.